Initial CiClops Website commit.
This commit contains the initial layout and templates for ciclops.cs.up.ac.za.
In addition a pre-alpha app is included that will serve as a web interface to
the Pleiades cluster.

Additional apps can be created later to interface with other components of
the CiClops project, such as the results database.

Signed-off-by: Bennie Leonard <bennie.leonard@gmail.com>